requirements.txt
altgraph==0.16.1
aniso8601==4.0.1
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
configparser==3.5.0
cryptography==2.3.1
docopt==0.6.2
enum34==1.1.6
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Mail==0.9.1
Flask-MySQL==1.4.0
Flask-RESTful==0.3.6
futures==3.2.0
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
ipaddress==1.0.22
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
macholib==1.11
MarkupSafe==1.0
mccabe==0.6.1
modulegraph==0.17
pipreqs==0.4.9
pkg-resources==0.0.0
py2app==0.18
pycparser==2.19
PyJWT==1.6.4
pylint==1.9.3
PyMySQL==0.9.2
pyparsing==2.2.2
python-dateutil==2.7.4
pytz==2018.6
requests==2.20.0
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.24
Werkzeug==0.14.1
wrapt==1.10.11
yarg==0.1.9
